UPDATE: _document.js added. Head moved into document
