+++ src/pages/_app.js
@@ -1,14 +1,13 @@
-import '../styles/globals.css'
-import '../styles/plugins/bootstrap.css'
-import '../styles/index.scss'
+import "../styles/globals.css";
+import "../styles/plugins/bootstrap.css";
+import "../styles/index.scss";
 import "react-toastify/dist/ReactToastify.css";
 import "animate.css";
 
-import { useState, useEffect } from 'react';
+import { useState, useEffect } from "react";
 import { IntlProvider } from "react-intl";
 import cx from "classnames";
 import lookie from "lookie";
-import Head from 'next/head'
 import { ToastContainer } from "react-toastify";
 
 import { Provider as LanguageProvider } from "../contexts/LanguageContext";
@@ -40,25 +39,11 @@ function MyApp({ Component, pageProps }) {
       <LanguageProvider lang={lang} setLang={handleChangeLang}>
         <ToastContainer />
         <div className={cx("App", { desktop: isDesktop })}>
-          <Head>
-            <script async src="https://www.googletagmanager.com/gtag/js?id=G-ZNDXNCMG6E" />
-            <script dangerouslySetInnerHTML={{
-              __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag() {dataLayer.push(arguments); }
-            gtag('js', new Date());
-            gtag('config', 'G-ZNDXNCMG6E');
-            `}} />
-            <link rel="shortcut icon" href="/favicon.svg" />
-            <meta name="description" content="Learn RegEx, step by step, zero to advanced. Playground and Cheatsheet." />
-          </Head>
           <Component {...pageProps} />
         </div>
       </LanguageProvider>
     </IntlProvider>
-  )
-
-
+  );
 }
 
-export default MyApp
+export default MyApp;
